refactor: move census and NCES domain constants from config/ to geo/ (SU#577)
census_registry.py and nces_constants.py are domain data consumed by
geo/ modules, not library configuration. Moved them to geo/ and left
deprecation shims in config/ (removal in v4.0.0).

Updated all 15 internal import sites in geo/ to use the new paths.
config/__init__.py re-exports from the new canonical locations.
